+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Current
+
+- GMX CCXT/FreqTrade integration
+- Add: Rewrite block timestamp cache to use DuckDB and Hypersync
+- Fix: Multiple RPC issue workarounds
+- Fix: Make `launch_anvil()` multi process locked, as likely Anvil has some internal conditions
+
 # 0.35
 
 - Add: Initial skeleton for GMX CCXT adapter